Allow straightforward compilation with Visual Studio 2017.
Visual Studio 2017 uses toolset v141 for now. It seems during the preview it was called v150? as it would have kept the pattern from previous versions.
Whatever the case was, "v$(VisualStudioVersion.Replace('.',''))" results in a lookup for v150, which is non-existent currently, so this replaces v150 with v141 which is the correct one.